@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.3.3'
+ruby '2.4.2'
 
 gem 'rails', '~> 5.2.0.beta2'
 gem 'jquery-rails'
