@@ -1,43 +1,91 @@
-set :application, "my_app_name"
-set :repo_url, "https://github.com/AndersonCamargo20/DeployOnDigitalOcean"
+# Change these
+server '157.230.81.239', port: 22, roles: [:web, :app, :db]
+
+set :repo_url,        'https://github.com/AndersonCamargo20/DeployOnDigitalOcean'
+set :application,     'DeployOnDigitalOcean'
+set :user,            'deploy'
+set :puma_threads,    [4, 16]
+set :puma_workers,    0
+
 
 set :user, 'deploy'
 set :deploy_to, "/var/www/awesome_bucket"
 
-append :linked_files, "config/database.yml", "config/storage.yml", "config/master.key"
-append :linked_dirs, "log", "tmp"
+#append :linked_files, "config/database.yml", "config/storage.yml", "config/master.key"
+#append :linked_dirs, "log", "tmp"
 
 
-#server "157.230.81.239", :primary => true
+# Don't change these unless you know what you're doing
+set :pty,             true
+set :use_sudo,        false
+set :stage,           :production
+set :deploy_via,      :remote_cache
+#set :deploy_to,       "/home/#{fetch(:user)}/apps/#{fetch(:application)}"
+set :puma_bind,       "unix://#{shared_path}/tmp/sockets/#{fetch(:application)}-puma.sock"
+set :puma_state,      "#{shared_path}/tmp/pids/puma.state"
+set :puma_pid,        "#{shared_path}/tmp/pids/puma.pid"
+set :puma_access_log, "#{release_path}/log/puma.error.log"
+set :puma_error_log,  "#{release_path}/log/puma.access.log"
+set :ssh_options,     { forward_agent: true, user: fetch(:user), keys: %w(~/.ssh/id_rsa.pub) }
+set :puma_preload_app, true
+set :puma_worker_timeout, nil
+set :puma_init_active_record, true  # Change to false when not using ActiveRecord
+## Defaults:
+# set :scm,           :git
+# set :branch,        :master
+# set :format,        :pretty
+set :log_level,     :debug
 set :keep_releases, 5
-set :migration_role, :app
 
-set :puma_pid, "#{shared_path}/tmp/pids/puma.pid"
-set :puma_bind, "unix://#{shared_path}/tmp/sockets/puma.sock"
-set :puma_access_log, "#{shared_path}/log/puma_access.log"
-set :puma_error_log, "#{shared_path}/log/puma_error.log"
-
-set :nginx_sites_available_path, "/etc/nginx/sites-available"
-set :nginx_sites_enabled_path, "/etc/nginx/sites-enabled"
-
-set :rvm_ruby_version, '2.3.3'
+## Linked Files & Directories (Default None):
+# set :linked_files, %w{config/database.yml}
+# set :linked_dirs,  %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system}
 
 namespace :puma do
-    desc 'Create Puma dirs'
-    task :create_dirs do
-      on roles(:app) do
-        execute "mkdir #{shared_path}/tmp/sockets -p"
-        execute "mkdir #{shared_path}/tmp/pids -p"
-      end
+  desc 'Create Directories for Puma Pids and Socket'
+  task :make_dirs do
+    on roles(:app) do
+      execute "mkdir #{shared_path}/tmp/sockets -p"
+      execute "mkdir #{shared_path}/tmp/pids -p"
     end
-  
-    desc "Restart Nginx"
-    task :nginx_restart do
-      on roles(:app) do
-        execute "sudo service nginx restart"
-      end
-    end
-  
-    before :start, :create_dirs
-    after :start, :nginx_restart
   end
+
+  before :start, :make_dirs
+end
+
+namespace :deploy do
+  desc "Make sure local git is in sync with remote."
+  task :check_revision do
+    on roles(:app) do
+      unless `git rev-parse HEAD` == `git rev-parse origin/master`
+        puts "WARNING: HEAD is not the same as origin/master"
+        puts "Run `git push` to sync changes."
+        exit
+      end
+    end
+  end
+
+  desc 'Initial Deploy'
+  task :initial do
+    on roles(:app) do
+      before 'deploy:restart', 'puma:start'
+      invoke 'deploy'
+    end
+  end
+
+  desc 'Restart application'
+  task :restart do
+    on roles(:app), in: :sequence, wait: 5 do
+      invoke 'puma:restart'
+    end
+  end
+
+  before :starting,     :check_revision
+  after  :finishing,    :compile_assets
+  after  :finishing,    :cleanup
+  after  :finishing,    :restart
+end
+
+# ps aux | grep puma    # Get puma pid
+# kill -s SIGUSR2 pid   # Restart puma
+# kill -s SIGTERM pid   # Stop puma
